+++ app/models/discussion_topic.rb
@@ -6,9 +6,9 @@ class DiscussionTopic < ActiveRecord::Base
   has_many :discussion_participants, inverse_of: :discussion_topic, dependent: :destroy
   has_many :participants, through: :discussion_participants, source: :user, class_name: 'User'
   has_many :discussion_replies, inverse_of: :discussion_topic, dependent: :destroy
+  has_many :notifications, as: :target
 
   def self.including(user)
     includes(:discussion_participants).where(discussion_participants: { user_id: user.id })
   end
-
 end