+++ config/initializers/event_stream_subscriptions.rb
@@ -50,6 +50,9 @@ Subscriptions.configure do
   add 'comment:created', stream_to_paper_channel
   add 'comment:destroyed', stream_to_everyone
 
+  add 'notification:created', stream_to_user
+  add 'notification:destroyed', stream_to_user
+
   add 'participation:created', stream_to_paper_channel
   add 'participation:destroyed', stream_to_everyone